# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Exercise 3.1'
  description: 'API - Web Development Environments'
...
paths:
...
  /users/getLastSearch:
    get:
      tags:
        - Users
      summary: Return the last search of the user
      responses:
        "200":
          description: data about the user's last search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LastSearch"
        "204":
          description: No search performed
        
        "401":
          description: User is not authorized
  /users/favoritePlayers:
    get:
      tags:
        - Users
      summary: Return the favorite players that were saved by user
      responses:
        "200":
          description: data about the user's favorites players returned 
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"
        "400":
          description: One of the players in the list is not found         
        "401":
          description: User is not authorized
    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerID:
                  type: string
                  description: player's ID which the user saved
                  example: 82808
      responses:
        "201":
          description: The player successfully saved as favorite
        "303":
          description: The player already in favorites team  
        "400":
          description: One of the argument is not specified  
        "401":
          description: The player is not exists
  /users/favoriteTeams:
    get:
      tags:
        - Users
      summary: Return the favorite teams that were saved by user
      responses:
        "200":
          description: data about the user's favorites teams returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "400":
          description: One of the players in the list is not found         
        "401":
          description: User is not authorized

    post:
      summary: save team in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamID:
                  type: string
                  description: team's ID which the user saved
                  example: 85
      responses:
        "201":
          description: The team successfully saved as favorite
        "303":
          description: The team already in favorites teams  
        "400":
          description: One of the argument is not specified  
        "401":
          description: The team is not exists
  /users/favoriteGames:
    get:
      tags:
        - Users
      summary: Return the favorite games that were saved by user
      responses:
        "200":
          description: data about the user's favorites games returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePreview"
        "400":
          description: One of the players in the list is not found         
        "401":
          description: user is not authorized

    post:
      summary: save game in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameID:
                  type: string
                  description: game's ID which the user saved
                  example: 1
      responses:
        "201":
          description: The game successfully saved as favorite
        "303":
          description: The game already in favorites teams  
        "400":
          description: One of the argument is not specified  
        "401":
          description: The game already accured || The game is not exists 
...
components:
  schemas:
    User:
      type: object
      properties:      
        userName: 
          type: string
          minLength: 3
          maxLength: 8
          pattern: ^[a-zA-Z]+$
          description: The UserName must conatins 3 to 8 letters!
          example: Shahar

        firstName:
          type: string
          description: First name of the user
          example: Shahar
          
        lastName:
          type: string
          example: Tal
          description: Last name of ther user

        country:
          type: string
          description: Uses's country
          example: Israel
          
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          pattern: ^[a-zA-Z0-9]*$
          example: Shahar123
          
        email:
          type: string 
          format: email
          example: ShaharAP@post.bgu.ac.il
          description: The email of the user
          
        image:
          type: string
          format: URL
          description: The user profile image
...
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust

        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png

        position:
          type: integer
          description: player's position in team
          example: 1

        team_name:
          type: string
          description: player's team name
          example: København
    CoachPreview:
      type: object
      properties:
        firsName:
          type: string
          description: The coach first name
        
        lastName:
          type: string
          description: The coach last name
          
        team_name:
          type: string
          description: The name of the team where the coach coaches
          example: Barcelona
        
        image:
          type: string
          format: URL
          description: The coach image
...
    TeamPreviewSearch:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København

        logo_path:
          type: string
          description: logo_path
    GamePreview:
      type: object
      properties:
        game_date:
          type: string
          description: Game date
          example: '2021-06-01'

        game_time:
          type: string
          description: Game time
          example: '16:30'

        home_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Home team
          
        away_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Away team  

        referee_name:
          items: 
            $ref: '#/components/schemas/Referee'
          description: Referee Name

        field:
          type: string 
          description: The field of the game        
...
    Coach:
      type: object
      properties:
        name:
          type: string
          description: The coach first name
        
        team_name:
          type: string
          description: The name of the team where the coach coaches
          example: Barcelona
        
        image:
          type: string
          format: URL
          description: The coach image

        common_name:
          type: string
          description: player's common name
          example: Sten  

        nationality:
          type: string
          description: player's nation
          example: Israel

        birthdate:
          type: string
          description: player's birthday
          example: Israel

        birthcountry:
          type: string
          description: player's country of birth
          example: Israel  
    eventType: 
      type: string 
      enum:
      - Goal
      - Offsife
      - Offense
      - Red Card
      - Yellow card
      - Injury
      - Subsitution
...
    Event:
      type: object
      properties:
        event_date:
          type: string
          description: The time of the event
          example: '01-01-2021'

        event_time:
          type: string
          description: The time of the event
          example: '16:30'
        
        minute:
          type: number
          description: The minute where the event accured
          example: 20
          
        event_description:
          type: string 
          description: Description of the event
          example: first goal of the game
          
        type:
          $ref: '#/components/schemas/eventType'
          example: 'Goal | Offsife | Offense | RedCard | YellowCard | Injury | Subsitution'
    Game:
      type: object
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - field
      properties:          
        game_time:
          type: string
          description: game date

        home_result:
          type: number
          description: The result of the game for the home team
        
        away_result:
          type: number
          description: The result of the game for the away team  

        home_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Home team
          
        away_team:
          items: 
            $ref: '#/components/schemas/Team'
          description: Away team  
          
        referee_name:
          items: 
            $ref: '#/components/schemas/Referee'
          description: Referee Name  

        field:
          type: string 
          description: The field of the game
          
        event_list:
          type: array
          minLength: 3 
          items:
            $ref: '#/components/schemas/Event' 
    Team:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København

        players:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/PlayerPreview"

        future_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        
        past_games:
          type: array
          items:
            $ref: '#/components/schemas/Game'    
            
        coach:
          type: object
          properties:
            items:
             $ref: '#/components/schemas/Coach'     
...
    Search:
      type: object
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerPreview'
          
        teames:
          type: array
          items:
            $ref: '#/components/schemas/TeamPreviewSearch'

        coaches:
          type: array
          items:
            $ref: '#/components/schemas/CoachPreview'  
...
    Referee:
      type: object
      properties:
        user:
          description: Referee details
          items:
            $ref: '#/components/schemas/User'

        training:
          type: string
          description: The training of the referee
          example: Superliga trainer
...
    LastSearch:
      type: object
      properties:
        query:
          type: string
          description: The query of the search
          example: /Sten

        queryResults:
           description: Show the result of the last search
           items:
            $ref: '#/components/schemas/Search'